[Event] Moving Dispatcher classes to a new Dispatcher namespace for better organization.
