+++ src/Valkyrja/Event/Dispatcher/Contract/Dispatcher.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Event\Contract;
+namespace Valkyrja\Event\Dispatcher\Contract;
 
 use Override;
 use Psr\EventDispatcher\EventDispatcherInterface;
