+++ tests/Unit/Application/ApplicationTest.php
@@ -47,8 +47,8 @@ use Valkyrja\Container\Data as ContainerData;
 use Valkyrja\Dispatcher\Contract\Dispatcher;
 use Valkyrja\Event\Collection\Contract\Collection as EventCollection;
 use Valkyrja\Event\Collector\Contract\Collector as EventCollector;
-use Valkyrja\Event\Contract\Dispatcher as EventDispatcher;
 use Valkyrja\Event\Data as EventData;
+use Valkyrja\Event\Dispatcher\Contract\Dispatcher as EventDispatcher;
 use Valkyrja\Http\Client\Contract\Client;
 use Valkyrja\Http\Client\GuzzleClient;
 use Valkyrja\Http\Client\LogClient;
