+++ tests/Unit/Event/Provider/ServiceProviderTest.php
@@ -21,9 +21,9 @@ use Valkyrja\Event\Collection\Collection;
 use Valkyrja\Event\Collection\Contract\Collection as CollectionContract;
 use Valkyrja\Event\Collector\AttributeCollector;
 use Valkyrja\Event\Collector\Contract\Collector;
-use Valkyrja\Event\Contract\Dispatcher as Contract;
 use Valkyrja\Event\Data;
-use Valkyrja\Event\Dispatcher;
+use Valkyrja\Event\Dispatcher\Contract\Dispatcher as Contract;
+use Valkyrja\Event\Dispatcher\Dispatcher;
 use Valkyrja\Event\Provider\ServiceProvider;
 use Valkyrja\Reflection\Contract\Reflection;
 use Valkyrja\Tests\Unit\Container\Provider\ServiceProviderTestCase;
